third_party/blink/web_tests/svg/dynamic-updates/SVGForeignObjectElement-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests
Test passed

Tests dynamic updates of the 'y' attribute of the SVGForeignObjectElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foreignObjectElement.getAttribute('y') is "-80"
PASS foreignObjectElement.getAttribute('y') is "80"
PASS successfullyParsed is true

TEST COMPLETE
